fix(xremap): Restore colon swap and bar mappings
I replaced them at f2528ca with Dygma Raise 2 but it turned out it does not work well on some cases, so I gave up that approach for now.

Goal:
I want to remap some keys when they are pressed with Shift.
I want to achieve that without key remapping software.

Approach using mechanical keyboard:
- Define a "pseudo-shift" layer which is almost same as the normal Shift key. like:
    - a -> A
    - b -> B
    - ...
- Change some key bindings on that layer. like:
    - normal: semicolon -> colon / pseudo: colon -> semicolon
    - normal: minus -> underscore / pseudo: minus -> bar

Problem:
That approach works almost fine.
But since the Shift key is mapped to a layer toggle for "pseudo-shift",
that key itself does not send `Shift` by it own anymore.
So for example, I cannot use a mouse with Shift.
I cannot select text by moving the text cursor by mouse with pressing Shift.
That is pain. And it cannot be solved with Dygma Raise.