feat(tmux): Consider outstanding active pane marker but don't change
- When I came back to the terminal, I want to quickly find the active pane.
- When I'm working in the terminal, I don't want the active pane to be
  too outstanding because it is annoying.

It may be ideal if the tmux can highlight the active pane only when
I focus the terminal, but it is not achievable only by tmux.

This option make the active pane more outstanding.
I like it but it is a bit annoying.
Therefore, this PR leave it as a comment and doesn't apply it.

```
set -g pane-border-format " #{pane_index}#{?pane_active, #[reverse] ,}#[default] "
```