feat(xremap): Experimentally bind Super-v to Ctrl-Shift-v
I changed the key mapping of paste in Alacritty in the previous commit
from `Ctrl-v` to `Ctrl-Shift-v`.
But I don't want to use a different mapping only in terminals.

So this commit changed the Xremap config of `Super-v`.
I think `Shift` is just ignored on pasting by `Ctrl-v` in most cases.
So this key binding allows me to use `Super-v` consistently.

- In normal apps, it works just like `Ctrl-v` .
- In Alacritty, it works as pasting thanks to `Shift`.