feat(git): Restore alias which does `git diff $(git merge-base A B) B`
Recently, I learned that the same thing can be done with `git diff A...B`.
Now I often use `git diff main...` when I want to see all the changed
made in the feature branch without seeing the changes made in the main
branch after branching.

but I also know that the meaning of `..` and `...` are different between
`git diff` and `git log`.
https://stackoverflow.com/questions/53120060/why-double-dot-vs-triple-dot-in-git-behaves-different-for-diff-and-log

And according to the link in the above page, one core Git maintainer says this is a mistake in design.
https://lore.kernel.org/git/CAEW0o+gYqWT5u-Tf8aDoMgXaf36Mb-XOApLNs4D+GMVLvsOjxg@mail.gmail.com/T/

> Please unlearn dot-dot and three-dots when using "git diff", which is not about ranges but about comparing two endpoints. If we were reinventing Git today from scratch, we would make "git diff A..B" an error. You can consider it a bug that the command accepts a range notation, but this will not change any time soon without a large fight to find and fix uses of the syntax in scripts by longtime Git users have written over the years.

I don't think dots notations will be changed or removed in the future,
but considering this point, it may be safer to use the `--merge-base`
option instead of three dots. Therefore, I restored the alias. But
thanks to the option, it is much simpler and flexible than the original
alias.