fix(git): Use "less -+F" as always-paging less
Overriding LESS no longer works since I moved configs for less to the lesskey file which is prioritized than the environment variable. It was also confusing since it needs to "specify the same options as the default LESS except for -F". Specyfing "-+F" to cancel the "-F" option is much clearer.