feat(vim/lsp): Add RuboCop LSP for Ruby files
Shopify/ruby-lsp is very heavy and often dos not work well in my environment. But RuboCop LSP launches quickly and work out of the box.

I can also use it for format-on-save and it works fine for small files,
but I didn't since it takes a few seconds on large files.